Changed Warden AI to assistant directory
